add a header to force bundle-uri advertisement despite rollout status
Summary: Sometimes we've got bundles and metadata for them ready to go for a repo, but the repo is not enabled via a JK. Such flag will make it much easier for us to do quick testing if a repo is good or not to be cloned via bundles. This will also allow us to clone repo without bundles even when they're enabled.

Reviewed By: RajivTS

Differential Revision: D74636133

fbshipit-source-id: 771a2a94855d45672fcfc459f265c9cf98c76f59