Fix a test that's recently become flaky
Summary:
D73667367 changed this test as a side effect of unrelated changes. In doing so, it became subtly depenent on the content of the table.

I don't actually understand how the query even works, it shouldn't be possible to mix aggregates and plain columns, especially with a GROUP BY.

The actual content of the table seems stable, so it's easier and probably safer to just include it all.

Differential Revision: D74582635

fbshipit-source-id: 3b20a5fafa370e63b494815d04f52819c071b309