Add tree aux data as part of a tree
Summary:
Modify Tree model so it contains a pointer to tree aux data. This helps to include the tree aux data from sapling backing store.

Note that this design has several flaws:
- The aux data will actually be stored side by side with the tree data so it bumps the actual size of the tree object thus decreasing the number of cached items from the tree cache. Since the overhead added by the tree aux data is fixed, the impact should be limited.
- If the aux data is later queried for the same tree, we can end up storing the same piece of tree aux data in the memory twice. A solution is to check the tree cache before checking the in-memory tree aux cache.

## Context
Currently we fetch TreeAuxData directly from the backing store without caching. We want to add cache support. The cache will be consist of the in-memory cache (in ObjectStore) as well as the local store.

Reviewed By: liubov-dmitrieva

Differential Revision: D73795236

fbshipit-source-id: 2f73fe758305d6e820e4d6f7c0e381d9ea4ea554