Add config flags and define counters
Summary:
Add configs to control the logic of how we want to prewarm the caches for tree aux data. When fetching a tree, the tree can be fetched from local store (within EdenFS), or from the sapling backing store.

If a tree is fetched from the local store and if it does not have a tree aux attached, we check `warmTreeAuxCacheIfTreeFromLocalStore` to decide if we want to trigger a tree aux request so the cache can be warmed.

If a tree is fetched from the backing store, we check `warmTreeAuxLocalCacheIfTreeFromBackStore` and `warmTreeAuxMemCacheIfTreeFromBackStore` to decide if we want to warm the in-memory cache and the local store cache respectively.

Add metrics to count the hits of cases above.

## Context
Currently we fetch TreeAuxData directly from the backing store without caching. We want to add cache support. The cache will be consist of the in-memory cache (in ObjectStore) as well as the local store.

Reviewed By: jdelliot

Differential Revision: D74187727

fbshipit-source-id: 8c9b6dfc5aaf4671047e9a8d739d721e491aaa1a