Include Tree aux data in the Tree response from sapling backing store
Summary:
Modify `fromRawTree` so it can populate the tree aux data when converting from a sapling tree into the eden tree. We check the returned digest_size of the tree aux data to determine if it's a valid aux data. We skip storing it if it's invalid.
Later we might want to force sapling to always return a valid tree aux data but for now let's go with the naive way.

## Context
Currently we fetch TreeAuxData directly from the backing store without caching. We want to add cache support. The cache will be consist of the in-memory cache (in ObjectStore) as well as the local store.

Reviewed By: jdelliot

Differential Revision: D73795248

fbshipit-source-id: a8524760a90780842b73c036df4218bcaaf030d5