Cache Tree aux when a tree is fetched with aux data available
Summary:
Prewarm the tree aux cache when a tree is fetched and it has tree aux data attached. In the previous diff we modified eden Tree so it includes a pointer to Tree aux data. The field will be populated if the tree fetched from sapling backing store has the tree aux data attached, which should always be the case.

There are two cases where we want to prewarm the cache:
- When a tree is fetched locally and it does not contain the tree aux, we trigger a tree aux query so the cache can be populated. This can be optimized since blocking the thread on a write lock
- When a tree is fetched

## Context
Currently we fetch TreeAuxData directly from the backing store without caching. We want to add cache support. The cache will be consist of the in-memory cache (in ObjectStore) as well as the local store.

Reviewed By: jdelliot

Differential Revision: D74188095

fbshipit-source-id: 99cbb6f9cc35ae7b20c6a496e37aab25d70c3f02