Log the error prefix trace to scuba
Summary:
We need some data in scuba to categorize sapling errors between user errors and actual concerning errors.
This is necessary to set-up slick metrics that are meaningful.

This diff is the simplest way to achieve this as we will be able to do some regex filtering on the errortrace_head.

Reviewed By: lmvasquezg

Differential Revision: D74586458

fbshipit-source-id: 931e94fa1e181f98f4185ade2e3b8e561b1b40cf