backingstore: be more selective logging prefetch activity
Summary:
Only log about prefetches if the walk depth is greater than 2. Otherwise, the event is too spammy to send to eden logs by default.

I used a macro for conditional levels because the level argument needs to be const, so it isn't otherwise convenient to choose the level at runtime.

Reviewed By: jdelliot

Differential Revision: D74707598

fbshipit-source-id: 4bbcce8e7d6a851da7d6212df9903e858411f685