Create new crate, remote-loader
Summary:
# This Diff

Begin to refactor remote loading bits into new crate - remote-loader.

# Next Steps

Continue to refactor the bits that are purely associated with fetching configurations from configerator. This will involve removing dependencies on Generater, ConfigSet and other Sapling specific types.

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: quark-zju

Differential Revision: D74726719

fbshipit-source-id: d8a5c91448702655911e2c3d11bee62137696c92