Refactor code from config/loader into config/fb/remote-loader
Summary:
# This Diff

Refactor code from config/loader into config/fb/remote-loader

# Next Steps

Continue to refactor the bits that are purely associated with fetching configurations from configerator. This will involve removing dependencies on Generater, ConfigSet and other Sapling specific types.

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: quark-zju

Differential Revision: D74728390

fbshipit-source-id: edaf0c6900de26979c9eb914598fc04e3326e361