client: generate Rust structs for scm usecases config
Summary:
# This diff

Generates some Rust structs which are needed to parse configerator configs after fetching them via graphql.

# Context

We'll use these structs to implement (de)serialization for scm usecases config that's fetched from configerator.

The usecases config will be used to determine whether a given use-case-id is valid, and if so, which limits should be applied to that use case

Reviewed By: jdelliot

Differential Revision: D74729986

fbshipit-source-id: 653afe4d3f564c9c1213ae6505a982fd2d636de4