Disable logging of retries
Summary:
These extra log entries are distracting from the main point of the test: no number of retries will make the request succeed.

On top of that though, I've been seeing flakiness from:

```
+  [WARN] Found error: Filenodes upload: Expected 2 responses, got 1, retrying attempt #0
```

That's weird that we're seeing mismatches (probably timeouts really), but again it's got nothing to do with this test.

Reviewed By: lmvasquezg

Differential Revision: D74802029

fbshipit-source-id: be3190a73d99a0913e4d81b4615a1191e49bc509