Add snapshot of current state of remote configs
Summary:
# This Diff

Add snapshot of current state of remote configs. For now, this has been done manually. We should add a more automate method in the future.

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: muirdm

Differential Revision: D74803774

fbshipit-source-id: 266602612cef1e9b27a936b24cb8b8e756ff1995