Add missing use case ids; update remote config snapshot
Summary:
# This Diff

Add missing use case ids; update remote config snapshot

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: MichaelCuevas

Differential Revision: D74804753

fbshipit-source-id: 3480cb01e4e310427a3a5d56b175ec365a444c15