Refactor use_case code + thrift types into mods
Summary:
# This Diff

Refactor use_case code + thrift types into mods.

# Next Steps

Continue to refactor the bits that are purely associated with fetching configurations from configerator. This will involve removing dependencies on Generater, ConfigSet and other Sapling specific types.

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: muirdm

Differential Revision: D74803772

fbshipit-source-id: 5736bb0e0607bf4c514abfc7918725db4e854c2b