Add update remote snapshot script and use to perform update
Summary:
# This Diff

The is script ensures (after a successful run) we have valid snapshot of the configerator state.

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: muirdm

Differential Revision: D74874869

fbshipit-source-id: 923dd54470d5cf567ae205f055a2bbb258fd8222