Load remote configs - not using it yet....
Summary:
# This Diff

Load remote configs - not using it yet....

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: muirdm

Differential Revision: D74808199

fbshipit-source-id: 7b11ec72fa3a7c662488517971524dd74213f1cd