Implement simplified helper methods for setting up http_client
Summary:
# This Diff

Implements some simplified helper methods - simplified in that they do not require Domain (Sapling construct) to use them - for setting up http_client.

# Next Steps

Continue to refactor the bits that are purely associated with fetching configurations from configerator. This will involve removing dependencies on Generater, ConfigSet and other Sapling specific types.

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: andreacampi

Differential Revision: D74803773

fbshipit-source-id: af6c17435fd01b94b349b6fb057747803179a009