Use queuing logger to ensure telemetry logging doesn't block the task
Summary: The default scuba client logger using scribe cat tries to immediately log the sample and returns only when the logging is complete. For the Eden thrift library, we expect to have a high throughput of telemetry events and we would like the logging to be non-blocking (i.e. happening on a separate thread). Recently, Muir implemented exactly that so let's use it for the Eden Thrift client library

Differential Revision: D74879825

fbshipit-source-id: 3060a90135e745051b7e650762c35cfcd54ac73d