Use remote configs for use cases with fallback to defaults
Summary:
# This Diff

Use remote configs for use cases with fallback to defaults

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: MichaelCuevas

Differential Revision: D74878450

fbshipit-source-id: 2b32d20c4098b0151b572489631856ef75577cee