Update remote config snapshot (includes flow and hack now)
Summary:
# This Diff

Update remote config snapshot (includes flow and hack now)

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: MichaelCuevas

Differential Revision: D74879445

fbshipit-source-id: 98ed05b80609e6839d53dfa39cf269f117036535