Integration test highlighting lack of protocol v1 support for Mononoke Git read path
Summary:
Since Git protocol V2 has been the norm and default for all Git client for close to a decade now, we decided to go ahead and implement only V2 in Mononoke Git since it was more flexible and offered options like shallow, filter, etc. However, Windows laptop users at Meta are stuck with a Git version that is more than a decade old. I tried updating the chef recipe to pick the latest Git version but that did not work.

Talking with members in the team, the consensus is that either we need to buckify and upgrade Git for everyone or we need to support protocol V1 for read path. Since Git buckification is a big project on its own requiring multiple months of effort and we want to get rid of Metagit before that, for now I have decided to go ahead with implementing V1 for the read path.

This diff includes an integration test that demonstrates that we do not support protocol V1 in Mononoke Git for the read path.

Reviewed By: YousefSalama

Differential Revision: D74995382

fbshipit-source-id: 296679c672edfb1326ceb3e80fba5f3023dd1d28