Take the set_intentionally_unmounted in a separate function
Summary:
## Context:
If the `intentionally-unmounted` flag is not set and we run graceful restart, there is a possibility that during the graceful restart the auto unmount recovery runs and recover the old eden instance. This is not expected and can destroy the graceful restart. To fix this issue, we need to set the `intentionally-unmounted` flag before graceful restart and clear it after restart finish.

## This diff:
Take `_set_intentionally_unmounted` and `_clear_intentionally_unmounted` functions into separate functions. Then they can get called from other places.

Reviewed By: genevievehelsel

Differential Revision: D74214224

fbshipit-source-id: 746c03402790ccb14d6faed09e1e51662b7bc3e4