turn off unmount recovery during graceful restart
Summary:
## Context:
If the `intentionally-unmounted` flag is not set and we run graceful restart, there is a possibility that during the graceful restart the auto unmount recovery runs and recover the old eden instance. This is not expected and can destroy the graceful restart. To fix this issue, we need to set the `intentionally-unmounted` flag before graceful restart and clear it after restart finish.

## This diff:
Set the flag before graceful restart and clear it after restart.

NOTE: If the graceful restart crash for any reason and a new eden start, it automatically clear the FLAG for new mounts. see here:
https://www.internalfb.com/code/fbsource/[6d0f557f832a]/fbcode/eden/fs/service/EdenServer.cpp?lines=1811

Reviewed By: genevievehelsel

Differential Revision: D74216806

fbshipit-source-id: 3fc0464f8f5088251ce3797069c40394fe7da10b