Add unit test for Graceful restart with too many inodes
Summary:
## Context:
EdenFS crash during graceful restart when the `TakeoveData` is large. One solution for this crash is to spilit the `TakeoverData` into 1 GB chunks, and send data over socket in chunks. More explanations and pseudocode is in this [post](https://fb.workplace.com/groups/1708850869939124/permalink/1952362862254589/)

## This diff:
Add a new unit test `manyInodes`
Build info for 10 million inodes. This exercises the code where we send `TakeoverData` which is larger than 1 GB (the maximum size of a single message over `UnixSocket`). The `TakeoverServer` should split the data into multiple chunks and send them in sequence.

Reviewed By: jdelliot

Differential Revision: D73958495

fbshipit-source-id: d45d24ddf94aa3fb6ccf3a1bf60360fd4886b3c9