add CHUNKED_MESSAGE protocol to version seven
Summary:
## Context:
EdenFS crash during graceful restart when the `TakeoveData` is large. One solution for this crash is to spilit the `TakeoverData` into 1 GB chunks, and send data over socket in chunks. More explanations and pseudocode is in this [post](https://fb.workplace.com/groups/1708850869939124/permalink/1952362862254589/)

## This diff:
Add the new protocol `TakeoverCapabilities::CHUNKED_MESSAGE` to version seven.
For now, this new protocol doesn't chunk the data. It only sends a first chunk message in the beginning then sends all the data in one chunk, and finally it sends a last chunk message.
In the next diffs, we will add new algorithms for chunking data in the server and reconstructing the chunked data in the client side.

Reviewed By: jdelliot

Differential Revision: D73411932

fbshipit-source-id: c91a2ecf0448eea8db77e7899e924b8a5b88f5b0