split and send the data in chunks
Summary:
## Context:
EdenFS crash during graceful restart when the `TakeoveData` is large. One solution for this crash is to spilit the `TakeoverData` into 1 GB chunks, and send data over socket in chunks. More explanations and pseudocode is in this [post](https://fb.workplace.com/groups/1708850869939124/permalink/1952362862254589/)

## This diff:
split the TakeoverData into chunks after serialization. Then Server sends the data chunks, and the Client receives the data in chunks, merges chunks together, and deserialize data.
The test shows that the new `TakeoverCapabilities::CHUNKED_MESSAGE` protocol can transfer data which are bigger than 2GB

NOTE: `UnixSocket::Message` contains `files` and `data`. We only split the `data` in multiple chunks after serialization. The `TakeoverServer` send `files` with the first data chunk in one message as the first data message. Then for the rest of the chunk messages they have data chunk but the `files` are empty. The `TakeoverClient` should consider this protocol in merging the data chunks and deserializing the message

Reviewed By: jdelliot

Differential Revision: D73559767

fbshipit-source-id: 305cdc5f20c243a87049e856228c7e01e32bb435