Add a config to control rollout of CHUNKED_MESSAGE protocol
Summary:
## Context:
EdenFS crash during graceful restart when the `TakeoveData` is large. One solution for this crash is to spilit the `TakeoverData` into 1 GB chunks, and send data over socket in chunks. More explanations and pseudocode is in this [post](https://fb.workplace.com/groups/1708850869939124/permalink/1952362862254589/)

## This diff:
Add a new config for controlling roll out of CHUNKED_MESSAGE protocol. This diff temporary adds `shouldChunkTakeoverData()` to `TakeoverHandler` then the `TakeoverServer` can access the `shouldChunkTakeoverData` value which is set in `EdenConfig`. The config value overrides the protocol capability. Therefore, when we set this config to false, the server won't chunk the data and the client will get the data in one chunk.
The plan is to open this config for the team to test it in prod, and then roll out the config to all. We can remove this config and all the added functions in this diff after roll out is completes to 100%

Reviewed By: jdelliot

Differential Revision: D75012879

fbshipit-source-id: ab88f2b7f58fc4cb7ef4ea4c590416a55f20a032