commit cloud: have the git servcie extend the SLAPI service
Summary:
We added the `git service` handler when we thought we'd have a separate API to serve commit cloud for git, now we have managed to use SLAPI, let's not have duplicated handlers.

This diff basically blocks endpoints that haven't been routed/tested but defaults to the SLAPI handler for anything else. The goal is to get rid of the git handler once we have ensured other flows work.

Reviewed By: RajivTS

Differential Revision: D74961893

fbshipit-source-id: 702c93e80557386a27378d737f9381a661f714ee