Set proxy socket path for use on vpnless
Summary:
# This Diff

Set proxy socket path for use on vpnless

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: liubov-dmitrieva

Differential Revision: D75172535

fbshipit-source-id: f2764194c2e8423a5b63c5ca6eefcd3ec5d33419