Update eden running version threshhold from 30 to 45 days
Summary:
As per the discussion in the document linked below, we will display EdenFS Health Report notifications for stale EdenFS versions only when the running version and installed version differ by more than 45 days. This change is part of the first proposal, which increases the gap in the Eden version from 30 days to 45 days.

https://docs.google.com/document/d/1fTxlasZfWv58iVhRr6Um3nSIOk4R0e4J6lxsCzG4W9E/edit?usp=sharing

Reviewed By: kavehahmadi60

Differential Revision: D74971406

fbshipit-source-id: ce01b40e991a14b6b67360f5e3988e9d62bbc784