Update remote config snapshot to include scm-eden-container
Summary:
# This Diff

Update remote config snapshot to include scm-eden-container

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: liubov-dmitrieva

Differential Revision: D75178708

fbshipit-source-id: b020820ccf9c037328f878b774469bf8a8482f94