walkdetector: add option for stricter threshold at shallower depths
Summary:
The cost of an overly aggressive walk detection is larger the shallower the walk is. For example, if a "random" file access pattern makes the detector briefly think there is a walk at the repo root, that walk will tend to stick around as the accesses continue, obscuring the real walk roots. This is not good since the "too high" of a walk root will cause wasted prefetch activity and in general is not accurate.

On the other hand, if we set the walk threshold to a high value, it will take signficantly longer to detect "happy" walks that are simply accessing every file under a root directory.

Address by adding a "lax depth" option to define what depth is no longer "too shallow". Walk roots that are shallower than the lax depth will get a (configurable) "strict" multiplier for their threshold. This makes it "harder" for a walk to bubble up to the repo root (or other high directories). Walk roots equal or deeper to the lax depth will use the base walk threshold value directly.

Reviewed By: zzl0

Differential Revision: D75188802

fbshipit-source-id: d5a72656a34cfbeab529aca8b8bb21db1e446719