walkdetector: add more nuanced handling of small and large dirs
Summary:
For very large dirs, we now will raise the walk threshold so it is at least 5% (by default) the size of the directory. For example, if a directory has 10k files and our walk threshold is 5, previously we would consider the directory "walked" when we saw 5 file loads. Now we would wait until we see 500.

For very small dirs where the directory size is smaller than the walk threshold, we previously would lower the walk threshold to 0 (i.e. any file read means the directory is walked). Now instead we cut the threshold in half until it is <= directory size. This won't make a difference for small thresholds, but if the walk threshold is larger then this approach will maintain a more conservative threshold even for smaller directories.

Reviewed By: zzl0

Differential Revision: D75188799

fbshipit-source-id: 109f52126952a66c0847d73cdaef08288efde626