Integration test highlighting incorrect Mononoke Git behaviour with tags
Summary:
There is an edge case in tag handling in Mononoke Git which was discovered as part of S520024 and was infact its root cause. Currently, you can make a repo inconsistent in Mononoke Git by following these steps:
- Create an annotated tag, e.g. refs/tag/testing pointing to an existing commit
- Push the annotated tag to server
- Locally, force (since Git doesn't allow this by default) update the tag to become a non-annotated tag pointing to some other commit
- Forcefully push this tag to the server (since Git doesn't allow it by default)

Now Mononoke Git would continue to have bonsai_tag_mapping for this tag (thinking its still an annotated tag) that points to an older commit but the tag's entry in the bookmarks table will reflect the new state of the world. The mismatch would lead to the repo being out of sync with Metagit.

Reviewed By: clara-9

Differential Revision: D75289503

fbshipit-source-id: c4a825374814d26f32ae9a7aa036230ac75b211f