backingstore: optimize TreeManifest usage in prefetch
Summary:
I realized it is very wasteful to use a fresh/empty TreeManifest for each prefetch operation. An empty TreeManifest will need to fetch the manifest blob and deserialize it for every tree it accesses during iteration, but prefetches tend to have good spacial locality, so the same trees are accessed over and over. Optimize by sharing a single TreeManifest betwewen prefetch operations. The TreeManifest is cheap to clone (as long as you aren't inserting new trees), and is thread safe, so should be perfect here.

When we have no active walks we drop the shared TreeManifest to clean up memory. It could be holding on to lots of trees.

Reviewed By: zzl0

Differential Revision: D75188794

fbshipit-source-id: d0f81d9754504988c319686c8c67b1246e7d52da