walkdetector: improve detection for slow walks
Summary:
Previously even if we were making progress towards advancing the depth of a walk, the walk would still get GC'd. Fix by bumping the walk's access time when a descendant makes progress.

I also tweaked the "should_advance_ancestor_walk" heuristic to allow the ancestor walk to be the direct parent of the current dir. This is convenient since the ancestor node's `advanced_children` map can track progress even after the descendants have been GC'd (i.e. we are slowly making progress advancing the depth). The other heuristics (like `maybe_merge_into_parent`) require all the children to be alive at the merge time, which might be unlikely if the walk threshold is large and the GC timeout is small.

Reviewed By: zzl0

Differential Revision: D75195214

fbshipit-source-id: 7553d5a760a7fd77040bbceeb46cc9ceac0087f4