backingstore: refactor prefetch to accept multiple walks
Summary:
I've noticed that our prefetch throughput is bad when there are a lot of small/shallow walks. We limit to 5 concurrent prefetches, and there is a lot of overhead for each prefetch (spawning a thread, etc), so we can't keep up when there are thousands of small walks at once.

I also realized that nothing about the prefetching is specific to only having a single walk, though. So, refactor the prefetch function to accept a list of walks. This way a lot of small walks can be prefetched together in a single "batch".

For now we only support batch directory prefetches. That is much more common in the workflows I was testing, and file content prefetching _does_ need to pass some per-walk metrics back to the walk detector (which would be hard to do with batches).

Reviewed By: zzl0

Differential Revision: D75188790

fbshipit-source-id: 5c454bc619740d7555005b40bd2e6ff91c2d1670