SaplingBackingStore: don't pass prefetched paths to walk detector
Summary:
Calls to `eden prefetch foo/**` are going to look like a walk to the walk detector, but it is not useful to detect them as walks and (double) prefetch. "Walk" refers to a serial walk, which precludes any kind of intentionally batched fetches.

Note that we do still track non-prefetch thrift calls such as readDir which could still be doing serial walking.

Reviewed By: zzl0

Differential Revision: D75258370

fbshipit-source-id: 15fde0a681da963b4f5e2a2ae80b77e888869500