fix invalidation in NFS GC
Summary:
GC invalidates inodes before deleting them from inode map. As there is no direct way to invalidates inodes on macOS we used a hack and invalidates inodes with `chmod`ing the inode. The `chmod`ing inodes only works on directories and not files.
When a directory invalidates, it invalidates all the files in the directory:
https://www.internalfb.com/code/fbsource/[6680b5af90f1]/fbcode/eden/fs/inodes/TreeInode.cpp?lines=4000
Therefore we don't need to invalidate files in the GC.

This diff limits macOS GC invalidations to only invalidate directories. When a directory gets invalidated, we can safely remove all the files inside that directory.
This change fixes the macOS GC

Reviewed By: genevievehelsel

Differential Revision: D75111045

fbshipit-source-id: 107280cdd403aab2abe78e91f2d139290697c654