github: add manylinux docker image
Summary:
I'd like to build a "portable" executable that runs on many Linux distros.
https://github.com/pypa/manylinux is a good starting point.

I chose `manylinux_2_34` instead of `manylinux_2_28` because it
has `libssl.so.3` and `libcrypto.so.3`. The latter has
`libssl.so.1` and `libcrypto.so.1`. Our binary will link to these
libraries (unlike `python`), modern Linux systems (2021+) have the
`.3` version.

Reviewed By: zzl0

Differential Revision: D75261379

fbshipit-source-id: adeccc1831d035a12e4b4760f3e31c0267382626