Introduce FilterContext for filtering fam events
Summary:
The diff introduces a new `FilterContext` struct in `file_access_monitor.rs` to enable filtering of fam events. This enables more efficient and flexible filtering of fam events, improving the overall performance and usability of the EdenFS commands.

### Example Use Cases

The `FilterContext` struct can be used to filter fam events based on path patterns and process IDs. For example:

```rust
let filter_context = FilterContext {
    path_pattern: Some("/path/to/filter".to_string()),
    pid: Some(1234),
};
```

It will be used in the next diffs.

Differential Revision: D74757619

fbshipit-source-id: cd4757725ca8df90b4bde9713401024746252902