Add support to focus on a single PID
Summary:
### Summary

This diff adds a new option to the `eden fam read` command to filter events by process ID (PID). The new `--pid` option allows users to specify a PID to focus on a single process, making it easier to analyze file access patterns for a specific process.
When "--pid" is provided, the final output will be counting how many times a path is accessed by the process.

Differential Revision: D74757618

fbshipit-source-id: fe9c4c2b83e76c7f62e08afbd9878b3948915b69