Add counter threshold
Summary:
The original concept of "top k" is useless since the user can always pipe the output into `head` to focus on the top N entries. So we are removing the argument "k".
But it would be useful to tell `fam read` to only show the entries with a counter larger than a specific value. This can improve the sorting performance and also make the output more clean.

Reviewed By: jdelliot

Differential Revision: D74824397

fbshipit-source-id: f295a1475df832c4c77e878caa24613e309db87f