Skip bootstrapping previous set of repos during service startup
Summary:
All our sharded services are configured to look at the previous stored state (either in file on disk or through SM integration with Zelos). In either case, we get the list of repos the task was supposed to have and then go ahead to load and bootstrap those repos. There are a couple of problems with this approach:
- The bootstrapping part is designed to be all or none, i.e. if we fail to setup even a single repo we give up the entire setup. This can have bad consequences in SEVs like S524005
- The bootstrapping part needs to finish before we can call our service "ready" for traffic. This means the bootstrapping adds a direct overhead when starting up tasks after restart or conveyor push
- It prevents the movement of shards across tasks. Sometimes while testing we restart tasks to move shards across the rest of the fleet forcing a rebalancing. This cannot happen if the task always loads up the previous saved state on restart

For these reasons, I have decided to opt out of loading the previous saved state on startup. Instead sharded services will now depend on shard manager to provide shards to be added to the task. This will ensure that we do not block the service from serving traffic immediately. It also ensures that we do not have the all-or-none semantics during service setup

Differential Revision: D75444000

fbshipit-source-id: eef6f4a344c13220bd60b031ed27b6c1319f43c9