config/loader: add logic to track new time shards
Summary:
The struct tracks "witness" (first saw) time of time shard rollouts.
With this we can detect potentially unsafe time shards, like:

   |13:00 ---------- 14:00|
    start              end
                    ^
            witness at 13:50  (potential cause: slow `conf ship`)

and automatically extend it to:

   |13:00 --------------------------------------- 15:30|
                    ^                      adjusted end
             witness (now)

to make the rollout safer. In case of real emergency, one can still change
the `start` time to, say, 1 year ago, to make it 99+% regardless of the
adjusted end.

Reviewed By: muirdm

Differential Revision: D74885135

fbshipit-source-id: 2c2fd5eb09bab45acacb10cfa7af75b523dc4ae1