prefetch_profiles: move killswitch check out of the EdenFSCheckout Object
Summary:
# Context
Predictive Prefetching[1] is a service offered by EdenFS that allows users or Sandcastle jobs to prefetch files based on historical access patterns. This feature offers a unique benefitsover other prefetching methods offered by EdenFS: it does not require pre-generated lists of directories to prefetch. Also, it is powered by one shared infrastructure for all use cases, where there are multiple backends driving normal prefetch profiles, and uses data with a couple of days lag, rather than "normal" prefetch profiles which use data which can be weeks or months stale due to flakey infrastructure.

The goal is to sunset the EdenFS owned "automatic generation" of normal Prefetch Profiles and offer Predictive Prefetch Profiles as an alternative - this will eliminate the need for EdenFS to maintain the flakey generation/compression infrastructure we wrote to autogenerate Prefetch Profiles using specs and Sandcastle jobs. It will also save on the Sandcastle capacity to generate these specs.

NOTE: EdenFS still will support fetching a "normal" Prefetch Profile, just if a team wants a periodically updated profile, they will be responsible for writing and maintaining their own infrastructure to generate and commit profiles to the repo

# This Diff
This moves changes the prefetching code to return if the `prefetch_profiles` function exits early due to the killswitch. When I was trying to show the behavior when the config is set off, the code prints like this:

```
Skipping Predictive Prefetch Profiles fetch due to global kill switch. This means prefetch-profiles.predictive-prefetching-enabled is not set in the EdenFS configs.
Started predictive prefetching in the background. Use 'eden trace hg' to monitor prefetch progress.
```

and with this change, we can skip printing the "started prefetching in the background" message. We should only print that if we're actually doing the prefetch.

Reviewed By: jdelliot

Differential Revision: D75489287

fbshipit-source-id: ee331cd637ed64e2fa18c1c2d5a908d70c2ed614