Provide appropriate error message when user tries to delete content ref in Mononoke Git server
Summary:
Mononoke Git currently doesn't support delete of refs pointing to content (i.e. trees and blobs). This is because we want to limit the interactions that users have with these content refs because they are an anti-pattern that we would like to avoid in our repo. The only content refs allowed are the ones that come from upstream where we have no choice. In the future, if there is demand we might consider supporting delete for such refs.

However, in the current version we accept the request for deleting such refs and return 200 OK making the client think that the deletion went through, when in reality it was a no-op on the server. This diff updates that logic so we have a valid error message that makes it clear to the user what the server did.

Reviewed By: YousefSalama

Differential Revision: D75519132

fbshipit-source-id: 92043adc9f0ebcd01d2c0aaf23a4c86a8da3a617