Log create/updates/deletes for git content refs to scribe through Monad command
Summary:
We now have caching implemented for `git_content_ref_mapping` that would get out of sync if we simply added/updated/deleted a content ref using the admin command. This is because the caches are configured to be updated based on events in the corresponding scribe category. In production, any such updates are logged to the scribe category which is why the caches get updated as expected.

This diff makes the necessary changes to ensure that we log such operations even when performed through monad command

Differential Revision: D75533983

fbshipit-source-id: 6f93c6af9b2f6ba8de38c8c2396d5d4e30792b18