use global concurrency semaphore
Summary: I want to limit concurrency of a process (git bundle generation) globally. I would like to have this with a semaphore, which blocks on acquire if there's no capacity to run, and wakes up when there is. I'm not sure if this is possible with Cerberus at the moment, so I might need to poll. Is it possible to do that prajay?

Reviewed By: RajivTS

Differential Revision: D75287629

fbshipit-source-id: 86c1a80c85dbe37cba7edce649d9a61f2c61cbb5