Make rpm.builder for fb-mononoke-admin
Summary:
The old release pipeline builds the rpm with this [packman.yml](https://www.internalfb.com/code/fbsource/fbcode/eden/mononoke/facebook/packman.yml)
That has been failing for a while, meaning that Mononoke devs have a stale version of `monad` on their dev servers.

I'm taking a few liberties compared to the original:
* Remove `mononoke_exec.sh` and `mononoke_exec_xrepo.sh` because these scripts were wrappers for the old admin command which has since been deprecated
* Strip the debug symbols because I strongly suspect that they are the cause for the [linking errors](https://www.internalfb.com/phabricator/paste/view/P1826282447) which prevent the old rpm from building. This also makes the rpm consistent with the [fbpkg](https://www.internalfb.com/code/fbsource/[4667e56e6b928f357d43f9428bb982b09b2fdcd3]/fbcode/eden/mononoke/facebook/fbpkg/BUCK?lines=28) where we also [strip the debug symbols](https://www.internalfb.com/code/fbsource/[4667e56e6b928f357d43f9428bb982b09b2fdcd3]/fbcode/eden/mononoke/facebook/fbpkg/lib.bzl?lines=36). See justifications for this choice in the diff description for D71816510.

This is a SEV followup for S523826 as we fixed blame derivation and we need this change in the hands of the team as soon as possible to avoid surprises.

This will need two follow-ups:
* Update the artifact ci config for fb-mononoke-admin to use this buck rule
* Setup a conveyor3 pipeline for fb-mononoke-admin

Reviewed By: andreacampi, RajivTS

Differential Revision: D75596800

fbshipit-source-id: 7f25502a8f96beef203c4d8c516daef5aeb4ce55