cli: change default value of predictive_prefetching_enabled killswitch
Summary:
# Context
Predictive Prefetching[1] is a service offered by EdenFS that allows users or Sandcastle jobs to prefetch files based on historical access patterns. This feature offers a unique benefitsover other prefetching methods offered by EdenFS: it does not require pre-generated lists of directories to prefetch. Also, it is powered by one shared infrastructure for all use cases, where there are multiple backends driving normal prefetch profiles, and uses data with a couple of days lag, rather than "normal" prefetch profiles which use data which can be weeks or months stale due to flakey infrastructure.

The goal is to sunset the EdenFS owned "automatic generation" of normal Prefetch Profiles and offer Predictive Prefetch Profiles as an alternative - this will eliminate the need for EdenFS to maintain the flakey generation/compression infrastructure we wrote to autogenerate Prefetch Profiles using specs and Sandcastle jobs. It will also save on the Sandcastle capacity to generate these specs.

NOTE: EdenFS still will support fetching a "normal" Prefetch Profile, just if a team wants a periodically updated profile, they will be responsible for writing and maintaining their own infrastructure to generate and commit profiles to the repo

# This Diff

This turns the default behavior for the killswitch to be "on"/"enabled" - there isnt a reason I can think of to turn this off by default, its really just meant to be an emergency killswitch. D75483297 turns the config on to `True` everywhere.

the normal "prefetch_profiles.prefetching_enabled" is defaulted to `true` too:
https://www.internalfb.com/code/fbsource/[341e6b7037e3e1e229983dc166b25941c8a90987]/fbcode/eden/fs/cli_rs/edenfs-client/src/checkout.rs?lines=1067-1069

Reviewed By: jdelliot

Differential Revision: D75489574

fbshipit-source-id: 8ce71a2766849fd1198acc7312e974cdc68dee4e