InferredCopyFrom: very basic derivation impl
Summary:
This diff added probably the simplest rename detection logic, e.g. if any of the new/changed file has the same content as a deleted one, then record them as renames.

It makes this type depend on `Fsnode` - we use it to query the "deleted file content" from parents.

Reviewed By: RajivTS

Differential Revision: D75254323

fbshipit-source-id: abdac0bbd60be90190b1c52581a87528bbf12510