cmdpy: set prefix to the exe dir when python is statically linked
Summary:
The intention of linking libpython statically is to ship a "portable" binary
without depending on a particular version of system libpython. By default,
the binary built this way will use an absolute path (ex.
`/opt/_internal/cpython-3.12.10` on manylinux) for Python stdlib.
We cannot demand such absolute path on a Linux system, and need to provide the
Python stdlib along with our binary. To do that, set the "prefix" to be
relative to the exe path.

Reviewed By: muirdm

Differential Revision: D75609445

fbshipit-source-id: e3d892ef5fba7fbc6af644da4704ce8f661fb1f1