Don't mock random ids in test
Summary: This test started failing because new code paths called randomId() more often than expected, since we mock randomId like 5 times. This is super fragile. The better way is to just find whatever ID the request was made with and have the tests make sure the matching ID is sent, which is what we're actually testing.

Reviewed By: quark-zju

Differential Revision: D75693282

fbshipit-source-id: affb8920d32e5b795ba8137e3557ca9e915e25e4