Transform fbcode/eden/fs
Summary:
This diff changes the struct initialization method from using positional arguments to keyword arguments, as indicated in the file title.

Background:

In thrift-python, user must use keyword arguments to initialize both mutable and immutable objects, meaning positional arguments are not allowed for this purpose. This behavior differs from py-deprecated, which permits both keyword and positional arguments for object initialization. To facilitate the migration from py-deprecated, we developed a tool that identifies all instances of struct initialization using positional arguments and converts them to keyword arguments. This will reduce the workload for users during the migration, and this diff is one of the outputs by using the tool.

Reviewed By: dtolnay

Differential Revision: D75805446

fbshipit-source-id: 3549aaf968fb411f7520e88541fce745a9e97a26