Ensure error messages for failed pushes do not exceed MAX_PACKETLINE_SIZE
Summary:
The response to git push from the server takes one of the following two forms:
- ok REF_NAME
- ng REF_NAME ERR_MSG

If the ERR_MSG in the above field becomes larger than 65,512 characters (i.e. MAX_PACKETLINE_SIZE), it technically becomes a new packet line for Git and Git ends up giving the following error message

```
error: invalid ref status from remote: SOME_RANDOM_STRING
```

We care about this because our SLIs track this as a server error that the server failed to resolve the push. However, its the error message formatting logic which is problematic. This diff fixes that.

Reviewed By: mzr

Differential Revision: D75868342

fbshipit-source-id: e434cd22063f386877c2545f5aeca1a8bcaeaa6e